+++ clash-ghc/CHANGELOG.md
@@ -9,6 +9,7 @@
 * New internal features:
   * [#821](https://github.com/clash-lang/clash-compiler/pull/821): Add `DebugTry`: print name of all tried transformations, even if they didn't succeed
   * [#856](https://github.com/clash-lang/clash-compiler/pull/856): Add `-fclash-debug-transformations`: only print debug info for specific transformations
+  * [#879](https://github.com/clash-lang/clash-compiler/pull/879): Add `-fclash-no-catch-errors`: don't try and catch errors in Clash compiler; let GHC handle it instead
 
 * Fixes issues:
   * [#810](https://github.com/clash-lang/clash-compiler/issues/810): Verilog backend now correctly specifies type of `BitVector 1`